api/requirements.txt
azure-functions==1.21.3
azure-cosmos==4.7.0
bcrypt==4.2.1
PyJWT==2.9.0
requests==2.32.3
